refactor: all

+++ lib/data/database/notes_database.dart
@@ -2,13 +2,11 @@ import 'dart:io';
 import 'package:sqflite/sqflite.dart';
 import 'package:path/path.dart';
 
-import '../models/note.dart';
-import '../models/tag.dart';
-import '../models/reminder.dart';
-import '../models/attachment.dart';
-import '../models/folder.dart';
-
-
+import '../models/note_model.dart';
+import '../models/tag_model.dart';
+import '../models/reminder_model.dart';
+import '../models/attachment_model.dart';
+import '../models/folder_model.dart';
 
 class NotesDatabase {
   // ================= SINGLETON =================
@@ -28,7 +26,7 @@ class NotesDatabase {
     if (Platform.environment.containsKey('FLUTTER_TEST')) {
       return openDatabase(
         inMemoryDatabasePath,
-        version: 9,
+        version: 10, // bumped version to add updatedAt
         onCreate: _createDB,
         onUpgrade: _upgradeDB,
       );
@@ -39,7 +37,7 @@ class NotesDatabase {
 
     return openDatabase(
       path,
-      version: 9,
+      version: 10, // bumped version to add updatedAt
       onCreate: _createDB,
       onUpgrade: _upgradeDB,
     );
@@ -53,8 +51,10 @@ class NotesDatabase {
         title TEXT,
         content TEXT,
         createdAt TEXT,
+        updatedAt TEXT,
         isPinned INTEGER DEFAULT 0,
-        isDeleted INTEGER DEFAULT 0
+        isDeleted INTEGER DEFAULT 0,
+        folderId INTEGER
       )
     ''');
 
@@ -83,9 +83,6 @@ class NotesDatabase {
       )
     ''');
 
-
-    
-
     await db.execute('''
       CREATE TABLE attachments (
         id INTEGER PRIMARY KEY AUTOINCREMENT,
@@ -95,7 +92,7 @@ class NotesDatabase {
       )
     ''');
 
-        await db.execute('''
+    await db.execute('''
       CREATE TABLE folders (
         id INTEGER PRIMARY KEY AUTOINCREMENT,
         name TEXT,
@@ -111,7 +108,6 @@ class NotesDatabase {
         PRIMARY KEY (folderId, noteId)
       )
     ''');
-
   }
 
   // ================= UPGRADE DB =================
@@ -173,7 +169,7 @@ class NotesDatabase {
       );
     }
 
-        if (oldVersion < 9) {
+    if (oldVersion < 9) {
       await db.execute('''
         CREATE TABLE IF NOT EXISTS folders (
           id INTEGER PRIMARY KEY AUTOINCREMENT,
@@ -192,49 +188,60 @@ class NotesDatabase {
       ''');
     }
 
+    if (oldVersion < 10) {
+      final columns = await db.rawQuery('PRAGMA table_info(notes)');
+      final hasUpdatedAt = columns.any((c) => c['name'] == 'updatedAt');
+
+      if (!hasUpdatedAt) {
+        await db.execute('ALTER TABLE notes ADD COLUMN updatedAt TEXT');
+      }
+
+      await db.execute(
+        'UPDATE notes SET updatedAt = createdAt WHERE updatedAt IS NULL',
+      );
+    }
   }
 
   // ================= NOTE =================
 
   /// CREATE
-  Future<int> createNote(Note note) async {
+  Future<int> createNote(NoteModel note) async {
     final db = await database;
     return await db.insert('notes', note.toMap());
   }
 
   /// FETCH NORMAL NOTES
-  Future<List<Note>> fetchNotes() async {
+  Future<List<NoteModel>> fetchNotes() async {
     final db = await database;
     final result = await db.query(
       'notes',
       where: 'isDeleted = 0',
-      orderBy: 'isPinned DESC, createdAt DESC',
+      orderBy: 'isPinned DESC',
     );
-    return result.map((e) => Note.fromMap(e)).toList();
+    return result.map((e) => NoteModel.fromMap(e)).toList();
   }
 
   /// FETCH TRASH
-  Future<List<Note>> fetchTrashNotes() async {
+  Future<List<NoteModel>> fetchTrashNotes() async {
     final db = await database;
     final result = await db.query(
       'notes',
       where: 'isDeleted = 1',
       orderBy: 'createdAt DESC',
     );
-    return result.map((e) => Note.fromMap(e)).toList();
+    return result.map((e) => NoteModel.fromMap(e)).toList();
   }
 
   /// GET BY ID
-  Future<Note?> getNoteById(int id) async {
+  Future<NoteModel?> getNoteById(int id) async {
     final db = await database;
-    final result =
-        await db.query('notes', where: 'id = ?', whereArgs: [id]);
-    if (result.isNotEmpty) return Note.fromMap(result.first);
+    final result = await db.query('notes', where: 'id = ?', whereArgs: [id]);
+    if (result.isNotEmpty) return NoteModel.fromMap(result.first);
     return null;
   }
 
   /// UPDATE
-  Future<int> updateNote(Note note) async {
+  Future<int> updateNote(NoteModel note) async {
     final db = await database;
     return await db.update(
       'notes',
@@ -249,7 +256,11 @@ class NotesDatabase {
     final db = await database;
     await db.update(
       'notes',
-      {'isDeleted': 1, 'isPinned': 0},
+      {
+        'isDeleted': 1,
+        'isPinned': 0,
+        'updatedAt': DateTime.now().toIso8601String(), // update timestamp
+      },
       where: 'id = ?',
       whereArgs: [id],
     );
@@ -260,7 +271,10 @@ class NotesDatabase {
     final db = await database;
     await db.update(
       'notes',
-      {'isDeleted': 0},
+      {
+        'isDeleted': 0,
+        'updatedAt': DateTime.now().toIso8601String(), // update timestamp
+      },
       where: 'id = ?',
       whereArgs: [id],
     );
@@ -270,19 +284,18 @@ class NotesDatabase {
   Future<void> deleteNotePermanently(int id) async {
     final db = await database;
 
-    await db.delete('reminders',
-        where: 'noteId = ?', whereArgs: [id]);
-    await db.delete('note_tags',
-        where: 'noteId = ?', whereArgs: [id]);
-    await db.delete('attachments',
-        where: 'noteId = ?', whereArgs: [id]);
+    await db.delete('reminders', where: 'noteId = ?', whereArgs: [id]);
+    await db.delete('note_tags', where: 'noteId = ?', whereArgs: [id]);
+    await db.delete('attachments', where: 'noteId = ?', whereArgs: [id]);
 
     await db.delete('notes', where: 'id = ?', whereArgs: [id]);
   }
 
   // ================= SEARCH =================
-  Future<List<Note>> searchNotesWithRange(
-      String keyword, DateTime fromDate) async {
+  Future<List<NoteModel>> searchNotesWithRange(
+    String keyword,
+    DateTime fromDate,
+  ) async {
     final db = await database;
     final result = await db.query(
       'notes',
@@ -291,51 +304,57 @@ class NotesDatabase {
         (title LIKE ? OR content LIKE ?) AND
         createdAt >= ?
       ''',
-      whereArgs: [
-        '%$keyword%',
-        '%$keyword%',
-        fromDate.toIso8601String(),
-      ],
+      whereArgs: ['%$keyword%', '%$keyword%', fromDate.toIso8601String()],
       orderBy: 'isPinned DESC, createdAt DESC',
     );
-    return result.map((e) => Note.fromMap(e)).toList();
+    return result.map((e) => NoteModel.fromMap(e)).toList();
   }
 
   // ================= TAG =================
-  Future<List<Note>> getNotesByTag(int tagId) async {
+  Future<List<NoteModel>> getNotesByTag(int tagId) async {
     final db = await database;
-    final result = await db.rawQuery('''
+    final result = await db.rawQuery(
+      '''
       SELECT notes.* FROM notes
       INNER JOIN note_tags ON notes.id = note_tags.noteId
       WHERE note_tags.tagId = ? AND notes.isDeleted = 0
       ORDER BY notes.isPinned DESC, notes.createdAt DESC
-    ''', [tagId]);
+    ''',
+      [tagId],
+    );
 
-    return result.map((e) => Note.fromMap(e)).toList();
+    return result.map((e) => NoteModel.fromMap(e)).toList();
   }
 
   Future<int> createTag(String name) async {
     final db = await database;
-    return await db.insert(
+    return await db.insert('tags', {
+      'name': name,
+    }, conflictAlgorithm: ConflictAlgorithm.ignore);
+  }
+
+  Future<void> renameTag(int tagId, String newName) async {
+    final db = await database;
+    await db.update(
       'tags',
-      {'name': name},
-      conflictAlgorithm: ConflictAlgorithm.ignore,
+      {'name': newName},
+      where: 'id = ?',
+      whereArgs: [tagId],
     );
   }
 
-  Future<List<Tag>> fetchTags() async {
+  Future<List<TagModel>> fetchTags() async {
     final db = await database;
     final result = await db.query('tags');
-    return result.map((e) => Tag.fromMap(e)).toList();
+    return result.map((e) => TagModel.fromMap(e)).toList();
   }
 
   Future<void> addTagToNote(int noteId, int tagId) async {
     final db = await database;
-    await db.insert(
-      'note_tags',
-      {'noteId': noteId, 'tagId': tagId},
-      conflictAlgorithm: ConflictAlgorithm.ignore,
-    );
+    await db.insert('note_tags', {
+      'noteId': noteId,
+      'tagId': tagId,
+    }, conflictAlgorithm: ConflictAlgorithm.ignore);
   }
 
   Future<void> removeTagFromNote(int noteId, int tagId) async {
@@ -347,24 +366,37 @@ class NotesDatabase {
     );
   }
 
-  Future<List<Tag>> getTagsOfNote(int noteId) async {
+  Future<void> deleteTag(int tagId) async {
     final db = await database;
-    final result = await db.rawQuery('''
+
+    await db.transaction((txn) async {
+      await txn.delete('note_tags', where: 'tagId = ?', whereArgs: [tagId]);
+
+      await txn.delete('tags', where: 'id = ?', whereArgs: [tagId]);
+    });
+  }
+
+  Future<List<TagModel>> getTagsOfNote(int noteId) async {
+    final db = await database;
+    final result = await db.rawQuery(
+      '''
       SELECT tags.* FROM tags
       INNER JOIN note_tags ON tags.id = note_tags.tagId
       WHERE note_tags.noteId = ?
-    ''', [noteId]);
+    ''',
+      [noteId],
+    );
 
-    return result.map((e) => Tag.fromMap(e)).toList();
+    return result.map((e) => TagModel.fromMap(e)).toList();
   }
 
   // ================= REMINDER =================
-  Future<int> createReminder(Reminder reminder) async {
+  Future<int> createReminder(ReminderModel reminder) async {
     final db = await database;
     return await db.insert('reminders', reminder.toMap());
   }
 
-  Future<List<Reminder>> getRemindersOfNote(int noteId) async {
+  Future<List<ReminderModel>> getRemindersOfNote(int noteId) async {
     final db = await database;
     final result = await db.query(
       'reminders',
@@ -372,34 +404,82 @@ class NotesDatabase {
       whereArgs: [noteId],
       orderBy: 'remindAt ASC',
     );
-    return result.map((e) => Reminder.fromMap(e)).toList();
+    return result.map((e) => ReminderModel.fromMap(e)).toList();
   }
 
-  Future<void> deleteReminder(int reminderId) async {
+  Future<int?> deleteReminder(int reminderId) async {
     final db = await database;
-    await db.delete(
+
+    final result = await db.query(
       'reminders',
       where: 'id = ?',
       whereArgs: [reminderId],
+      limit: 1,
+    );
+
+    if (result.isEmpty) return null;
+
+    final notificationId = result.first['notificationId'] as int;
+
+    await db.delete('reminders', where: 'id = ?', whereArgs: [reminderId]);
+    return notificationId;
+  }
+
+  Future<void> markReminderDone(int id) async {
+    final db = await database;
+    await db.update(
+      'reminders',
+      {'isDone': 1},
+      where: 'id = ?',
+      whereArgs: [id],
     );
   }
 
   // ================= ATTACHMENT =================
-  Future<void> addAttachment(Attachment attachment) async {
+  Future<void> addAttachment(AttachmentModel attachment) async {
     final db = await database;
     await db.insert('attachments', attachment.toMap());
   }
 
-  Future<List<Attachment>> getAttachmentsOfNote(int noteId) async {
+  Future<List<AttachmentModel>> getAttachmentsOfNote(int noteId) async {
     final db = await database;
-    final result =
-        await db.query('attachments', where: 'noteId = ?', whereArgs: [noteId]);
-    return result.map((e) => Attachment.fromMap(e)).toList();
+    final result = await db.query(
+      'attachments',
+      where: 'noteId = ?',
+      whereArgs: [noteId],
+    );
+    return result.map((e) => AttachmentModel.fromMap(e)).toList();
   }
 
   Future<void> deleteAttachment(int id) async {
     final db = await database;
+
+    final result = await db.query(
+      'attachments',
+      where: 'id = ?',
+      whereArgs: [id],
+      limit: 1,
+    );
+
+    if (result.isEmpty) return;
+
+    final filePath = result.first['filePath'] as String;
+
     await db.delete('attachments', where: 'id = ?', whereArgs: [id]);
+
+    final file = File(filePath);
+    if (await file.exists()) {
+      await file.delete();
+    }
+  }
+
+  Future<int> countAttachmentsOfNote(int noteId) async {
+    final db = await database;
+    final result = await db.rawQuery(
+      'SELECT COUNT(*) as count FROM attachments WHERE noteId = ?',
+      [noteId],
+    );
+    return Sqflite.firstIntValue(result) ?? 0;
   }
 
   // ================= CLOSE =================
@@ -409,20 +489,19 @@ class NotesDatabase {
   }
   // ================= FOLDER =================
 
-  Future<int> createFolder(Folder folder) async {
+  Future<int> createFolder(FolderModel folder) async {
     final db = await database;
     return await db.insert('folders', folder.toMap());
   }
 
-  Future<List<Folder>> fetchFolders() async {
+  Future<List<FolderModel>> fetchFolders() async {
     final db = await database;
     final result = await db.query(
       'folders',
       orderBy: 'color ASC, createdAt DESC',
     );
-    return result.map((e) => Folder.fromMap(e)).toList();
+    return result.map((e) => FolderModel.fromMap(e)).toList();
   }
-
 
   Future<int> countNotesInFolder(int folderId) async {
     final db = await database;
@@ -436,15 +515,14 @@ class NotesDatabase {
 
   Future<void> addNoteToFolder(int folderId, int noteId) async {
     final db = await database;
-    await db.insert(
-      'folder_notes',
-      {'folderId': folderId, 'noteId': noteId},
-      conflictAlgorithm: ConflictAlgorithm.ignore,
-    );
+    await db.insert('folder_notes', {
+      'folderId': folderId,
+      'noteId': noteId,
+    }, conflictAlgorithm: ConflictAlgorithm.ignore);
   }
 
   Future<void> deleteFolder(int folderId) async {
-  final db = await database;
+    final db = await database;
 
     // Xóa liên kết note – folder trước
     await db.delete(
@@ -454,44 +532,49 @@ class NotesDatabase {
     );
 
     // Xóa folder
-    await db.delete(
-      'folders',
-      where: 'id = ?',
-      whereArgs: [folderId],
-    );
+    await db.delete('folders', where: 'id = ?', whereArgs: [folderId]);
   }
   // ================= FOLDER NOTES =================
 
-  Future<List<Note>> fetchNotesByFolder(int folderId) async {
+  Future<List<NoteModel>> fetchNotesByFolder(int folderId) async {
     final db = await database;
 
-    final result = await db.rawQuery('''
+    final result = await db.rawQuery(
+      '''
       SELECT notes.* FROM notes
       INNER JOIN folder_notes
         ON notes.id = folder_notes.noteId
       WHERE folder_notes.folderId = ?
         AND notes.isDeleted = 0
       ORDER BY notes.isPinned DESC, notes.createdAt DESC
-    ''', [folderId]);
+    ''',
+      [folderId],
+    );
 
-    return result.map((e) => Note.fromMap(e)).toList();
+    return result.map((e) => NoteModel.fromMap(e)).toList();
   }
 
   Future<void> linkNoteToFolder(int noteId, int folderId) async {
     final db = await database;
 
-    await db.insert(
-      'folder_notes',
-      {
-        'noteId': noteId,
-        'folderId': folderId,
-      },
-      conflictAlgorithm: ConflictAlgorithm.ignore,
+    await db.insert('folder_notes', {
+      'noteId': noteId,
+      'folderId': folderId,
+    }, conflictAlgorithm: ConflictAlgorithm.ignore);
+  }
+
+  Future<void> renameFolder(int id, String name) async {
+    final db = await database;
+    await db.update(
+      'folders',
+      {'name': name},
+      where: 'id = ?',
+      whereArgs: [id],
     );
   }
 
   // ================= HOME NOTES (KHÔNG THUỘC FOLDER) =================
-  Future<List<Note>> fetchNotesWithoutFolder() async {
+  Future<List<NoteModel>> fetchNotesWithoutFolder() async {
     final db = await database;
 
     final result = await db.rawQuery('''
@@ -503,7 +586,7 @@ class NotesDatabase {
       ORDER BY isPinned DESC, createdAt DESC
     ''');
 
-    return result.map((e) => Note.fromMap(e)).toList();
+    return result.map((e) => NoteModel.fromMap(e)).toList();
   }
 
   Future<void> updateFolderName(int id, String name) async {
@@ -516,10 +599,10 @@ class NotesDatabase {
     );
   }
 
-Future<Map<int, int>> countNotesByFolder() async {
-  final db = await database;
+  Future<Map<int, int>> countNotesByFolder() async {
+    final db = await database;
 
-  final result = await db.rawQuery('''
+    final result = await db.rawQuery('''
     SELECT fn.folderId, COUNT(fn.noteId) as total
     FROM folder_notes fn
     INNER JOIN notes n ON n.id = fn.noteId
@@ -527,45 +610,42 @@ Future<Map<int, int>> countNotesByFolder() async {
     GROUP BY fn.folderId
   ''');
 
-  final Map<int, int> map = {};
-  for (final row in result) {
-    map[row['folderId'] as int] = row['total'] as int;
+    final Map<int, int> map = {};
+    for (final row in result) {
+      map[row['folderId'] as int] = row['total'] as int;
+    }
+
+    return map;
   }
 
-  return map;
-}
+  Future<void> moveNoteToFolder({
+    required int noteId,
+    int? folderId, // null = đưa ra ngoài folder
+  }) async {
+    final db = await database;
 
-Future<void> moveNoteToFolder({
-  required int noteId,
-  int? folderId, // null = đưa ra ngoài folder
-}) async {
-  final db = await database;
+    // 🔥 Xóa liên kết cũ
+    await db.delete('folder_notes', where: 'noteId = ?', whereArgs: [noteId]);
 
-  // 🔥 Xóa liên kết cũ
-  await db.delete(
-    'folder_notes',
-    where: 'noteId = ?',
-    whereArgs: [noteId],
-  );
-
-  // 🔥 Nếu có folder mới → thêm lại
-  if (folderId != null) {
-    await db.insert(
-      'folder_notes',
-      {
+    // 🔥 Nếu có folder mới → thêm lại
+    if (folderId != null) {
+      await db.insert('folder_notes', {
         'noteId': noteId,
         'folderId': folderId,
-      },
-      conflictAlgorithm: ConflictAlgorithm.ignore,
-    );
+      }, conflictAlgorithm: ConflictAlgorithm.ignore);
+    }
   }
-}
 
+  Future<List<Map<String, dynamic>>> getNotesWithoutFolder() async {
+    final db = await database;
 
-
-
-
-
-
-  
+    return await db.rawQuery('''
+    SELECT * FROM notes
+    WHERE isDeleted = 0
+      AND id NOT IN (
+        SELECT noteId FROM folder_notes
+      )
+    ORDER BY isPinned DESC, createdAt DESC
+  ''');
+  }
 }
